+++ service/src/main/java/uk/nhs/adaptors/gp2gp/common/exception/RetryLimitReachedException.java
@@ -1,0 +1,7 @@
+package uk.nhs.adaptors.gp2gp.common.exception;
+
+public class RetryLimitReachedException extends RuntimeException {
+    public RetryLimitReachedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}