+++ service/src/main/java/uk/nhs/adaptors/gp2gp/common/service/WebClientFilterService.java
@@ -1,10 +1,12 @@
 package uk.nhs.adaptors.gp2gp.common.service;
 
 import static java.util.regex.Pattern.CASE_INSENSITIVE;
+
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeoutException;
 import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -20,11 +22,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
+import reactor.util.retry.Retry;
+import uk.nhs.adaptors.gp2gp.common.configuration.WebClientConfiguration;
 import uk.nhs.adaptors.gp2gp.common.exception.MaximumExternalAttachmentsException;
+import uk.nhs.adaptors.gp2gp.common.exception.RetryLimitReachedException;
 import uk.nhs.adaptors.gp2gp.gpc.exception.EhrRequestException;
 import uk.nhs.adaptors.gp2gp.gpc.exception.GpConnectException;
 import uk.nhs.adaptors.gp2gp.gpc.exception.GpConnectInvalidException;
 import uk.nhs.adaptors.gp2gp.gpc.exception.GpConnectNotFoundException;
+import uk.nhs.adaptors.gp2gp.gpc.exception.GpcServerErrorException;
 import uk.nhs.adaptors.gp2gp.mhs.InvalidOutboundMessageException;
 import uk.nhs.adaptors.gp2gp.mhs.exception.MhsServerErrorException;
 
@@ -50,6 +56,12 @@ public class WebClientFilterService {
     private static final Map<RequestType, Function<String, Exception>> REQUEST_TYPE_TO_EXCEPTION_MAP = Map.of(
             RequestType.GPC, GpConnectException::new);
 
+    private static final List<Class<? extends Exception>> RETRYABLE_EXCEPTIONS = List.of(
+        TimeoutException.class,
+        MhsServerErrorException.class,
+        GpcServerErrorException.class
+    );
+
     public enum RequestType {
         GPC, MHS_OUTBOUND
     }
@@ -57,12 +69,15 @@ public class WebClientFilterService {
     public static void addWebClientFilters(
             List<ExchangeFilterFunction> filters,
             RequestType requestType,
-            HttpStatus expectedSuccessHttpStatus) {
+            HttpStatus expectedSuccessHttpStatus,
+            WebClientConfiguration clientConfiguration) {
 
         // filters are executed in reversed order
+        filters.add(retryPolicy(requestType, clientConfiguration));
         filters.add(errorHandling(requestType, expectedSuccessHttpStatus));
         filters.add(logRequest());
         filters.add(logResponse());
+        filters.add(timeout(clientConfiguration));
         // this will be executed as the first one - always needs to come first
         filters.add(mdc());
     }
@@ -79,16 +94,17 @@ public class WebClientFilterService {
 
     private static ExchangeFilterFunction errorHandling(RequestType requestType, HttpStatus httpStatus) {
         return ExchangeFilterFunction.ofResponseProcessor(clientResponse -> {
+
             var statusCode = clientResponse.statusCode();
             if (statusCode.equals(httpStatus)) {
                 LOGGER.info(requestType + " request successful status_code: {}", statusCode.value());
                 return Mono.just(clientResponse);
             }
+            if (statusCode.is5xxServerError()) {
+                return getInternalServerErrorException(clientResponse, requestType);
+            }
             if (requestType.equals(RequestType.GPC)) {
                 return getErrorException(clientResponse, requestType);
-            }
-            if (requestType.equals(RequestType.MHS_OUTBOUND) && statusCode.is5xxServerError()) {
-                return Mono.error(new MhsServerErrorException("MHS responded with status code " + statusCode.value()));
             }
             if (requestType.equals(RequestType.MHS_OUTBOUND) && statusCode.value() == BAD_REQUEST.value()) {
 
@@ -155,11 +171,6 @@ public class WebClientFilterService {
                     return NACK_ERROR_18;
                 }
                 break;
-            case INTERNAL_SERVER_ERROR:
-                if (codes.contains(INTERNAL_SERVER_ERROR_STATUS)) {
-                    return NACK_ERROR_20;
-                }
-                break;
             case UNPROCESSABLE_ENTITY:
                 if (codes.contains(INVALID_RESOURCE_STATUS)
                         || codes.contains(BAD_REQUEST_STATUS)
@@ -223,5 +234,51 @@ public class WebClientFilterService {
             }
             return Mono.just(response);
         });
+    }
+
+    private static Mono<ClientResponse> getInternalServerErrorException(ClientResponse clientResponse, RequestType requestType) {
+
+        if (requestType.equals(RequestType.MHS_OUTBOUND)) {
+            return Mono.error(new MhsServerErrorException(
+                String.format(REQUEST_EXCEPTION_MESSAGE, requestType, clientResponse.statusCode()))
+            );
+        }
+
+        return clientResponse.toEntity(String.class).flatMap(response -> {
+            String exceptionMessage;
+
+            if (response.hasBody()) {
+                exceptionMessage = String.format(REQUEST_EXCEPTION_MESSAGE, requestType, response.getBody());
+            } else {
+                exceptionMessage = String.format(REQUEST_EXCEPTION_MESSAGE, requestType, clientResponse.statusCode());
+            }
+
+            return Mono.error(new GpcServerErrorException(exceptionMessage));
+        });
+    }
+
+    private static ExchangeFilterFunction retryPolicy(RequestType requestType, WebClientConfiguration clientConfiguration) {
+        return (request, next) ->
+            next.exchange(request)
+                .retryWhen(getRetryPolicyByRequestType(requestType, clientConfiguration));
+    }
+
+    private static ExchangeFilterFunction timeout(WebClientConfiguration clientConfiguration) {
+        return (request, next) -> next.exchange(request).timeout(clientConfiguration.getTimeout());
+    }
+
+    private static Retry getRetryPolicyByRequestType(RequestType type, WebClientConfiguration clientConfiguration) {
+        return Retry
+            .backoff(clientConfiguration.getMaxBackoffAttempts(), clientConfiguration.getMinBackOff())
+            .filter(exception -> RETRYABLE_EXCEPTIONS.contains(exception.getClass()))
+            .doBeforeRetry(retrySignal ->
+                LOGGER.info("Request to `{}` failed, retrying request {}/{}",
+                    type, retrySignal.totalRetries() + 1, clientConfiguration.getMaxBackoffAttempts()))
+            .onRetryExhaustedThrow((retryBackoffSpec, retrySignal) ->
+                new RetryLimitReachedException(String.format("Retries exhausted: %s/%s",
+                        retrySignal.totalRetries(), clientConfiguration.getMaxBackoffAttempts()), retrySignal.failure()
+                )
+            );
+
     }
 }